requirements.txt
requests>=2.28.0
python-dotenv>=0.20.0
google-generativeai>=0.3.0
pytest>=7.0.0
pyyaml>=6.0